move autoscroll point
